requirements.txt
EbookLib
html2text
python-dotenv
request
llama-index
rich
openai
etils[etree]
tqdm
google-generativeai
pydantic